feat(subscription): implementation for control access with qr         - generation of qr when the user get a subcription        - return information about subscription of the one user

+++ src/main/java/com/GymManager/Backend/persistence/projections/AllDataAboutUser.java
@@ -7,11 +7,11 @@ import java.time.LocalDateTime;
 
 public interface AllDataAboutUser {
     // personal info
-    Integer getId();
+    Long getId();
     String getFullName();
     String getDni();
     LocalDate getDateOfBirth();
-    Integer getPhone();
+    Long getPhone();
     String getEmail();
     String getGender();
     LocalDateTime getCreateDate();
